% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Versão tidy dos dados "cpi_raw"}
\format{Data frame (tibble) com 2790 linhas e 3 colunas

\describe{
\item{Pais}{Pais.}
\item{Ano}{Ano.}
\item{CPI}{Índice de Percepção de Corrupção. O índice varia entre 0 e 100, onde quando mais próximo de 100 menor é a percepção de corrupção.}
}}
\usage{
cpi
}
\description{
Versão tidy dos dados "cpi_raw"
}
\keyword{datasets}
