% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_raw.R
\docType{data}
\name{cpi_raw}
\alias{cpi_raw}
\title{Dados do índice de percepção de corrupção}
\format{Data frame (tibble) com 186 linhas e 23 colunas}
\source{
\ref{http://knoema.com/CPI2014/corruption-perceptions-index-2016}{International Transparency}
}
\usage{
cpi_raw
}
\description{
Dados do índice de percepção de corrupção
}
\keyword{datasets}
