% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set1.R
\docType{data}
\name{data_set1}
\alias{data_set1}
\title{Data set que utilizada como exemplo no livro}
\format{Data frame (tibble) com 1184 linhas e 9 colunas

\describe{
\item{Pais}{País.}
\item{Ano}{Ano.}
\item{Regiao}{Região no mundo onde o Pais se encontra.}
\item{PerCap_dolar}{Pib Per Capita em dólares constantes de 2011.}
\item{PerCap_growth}{Variação percentual do Pib Per Capita em relaçao ao ano anterior.}
\item{Inflacao}{Variação percentual do principal índice de preços ao consumidor do pais.}
\item{CPI}{Índice de Percepção de Corrupção. O índice varia entre 0 e 100, onde quando mais próximo de 100 menor é a percepção de corrupção.}
\item{PFI}{Índice de Liberdade de Imprensa. O índice varia entre 0 e 100, onde quando mais próximo de 100 menor é a liberdade de imprensa.}
\item{PFI}{Índice de democracia. O pindice varia entre 0 e 10 onde quanto mais próximo de 10 mais democrático é o pais.}
}}
\usage{
data_set1
}
\description{
Data set que utilizada como exemplo no livro
}
\keyword{datasets}
