% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\docType{data}
\name{di}
\alias{di}
\title{Versão tidy dos dados "di_raw"}
\format{Data frame (tibble) com 1320 linhas e 3 colunas

\describe{
\item{Pais}{Pais.}
\item{Ano}{Ano.}
\item{PFI}{Índice de democracia. O pindice varia entre 0 e 10 onde quanto mais próximo de 10 mais democrático é o pais.}
}}
\usage{
di
}
\description{
Versão tidy dos dados "di_raw"
}
\keyword{datasets}
