% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_raw.R
\docType{data}
\name{di_raw}
\alias{di_raw}
\title{Dados do índice de democracia}
\format{Data frame (tibble) com 165 linhas e 16 colunas}
\source{
\ref{http://knoema.com/QOGISD2017/quality-of-government-institute-standard-dataset-version-january-2017}{University of Gothenburg}
}
\usage{
di_raw
}
\description{
Dados do índice de democracia
}
\keyword{datasets}
