% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfi.R
\docType{data}
\name{pfi}
\alias{pfi}
\title{Versão tidy dos dados "pfi_raw"}
\format{Data frame (tibble) com 2976 linhas e 3 colunas

\describe{
\item{Pais}{Pais.}
\item{Ano}{Ano.}
\item{PFI}{Índice de Liberdade de Imprensa. O índice varia entre 0 e 100, onde quando mais próximo de 100 menor é a liberdade de imprensa.}
}}
\usage{
pfi
}
\description{
Versão tidy dos dados "pfi_raw"
}
\keyword{datasets}
