% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfi_raw.R
\docType{data}
\name{pfi_raw}
\alias{pfi_raw}
\title{Dados do índice de liberdade de imprensa}
\format{Data frame (tibble) com 186 linhas e 23 colunas}
\source{
\ref{http://knoema.com/GSPFI2016/world-press-freedom-index-2017}{Reporters Without Borders}
}
\usage{
pfi_raw
}
\description{
Dados do índice de liberdade de imprensa
}
\keyword{datasets}
