% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdi_raw.R
\docType{data}
\name{wdi_raw}
\alias{wdi_raw}
\title{Subset do world development indicators}
\format{Data frame (tibble) com 723 linhas e 72 colunas}
\source{
\ref{http://knoema.com/WBWDIGDF2017Sep/world-development-indicators-wdi}{Banco Mundial}
}
\usage{
wdi_raw
}
\description{
Variáveis selecionadas do World Development Indicators sem
nenhum tipo de manipulação de dados.
}
\keyword{datasets}
